long 타입 int 로 수정

+++ api/src/main/java/com/frontier/records/api/music/model/LikeMoment.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 public class LikeMoment {
 
-    private long moment;
+    private int moment;
     private int likeCount;
 }
